Use loop to run load-test instead of cronjob
Instead of using a cronjob to run run-llm-load-test.py, now
use a loop. This is because of storage limitations. If we used
a cronjob the pvc would need to be readwritemany. The storage
I am using is readwriteonce so instead of using cronjob the
application will run as a container in the same pod as the
exporter.

Signed-off-by: Isaiah Stapleton <istaplet@redhat.com>